refactor(fsd): lift firebase/auth out of model, extract useStandingsScreen + view UI
Per .cursorrules §2/§3, models should not import firebase/* directly and
non-trivial pages should hoist orchestration into a feature-owned screen
hook. Two structural smells addressed:

1. useSplashSignUp imported `getAdditionalUserInfo` from `firebase/auth`
   directly. Moved into `splashAuthApi.signInWithGoogle`, which now resolves
   to `{ isNewUser }`. Added an ESLint guardrail blocking direct
   `firebase/{firestore,auth,functions,storage,app-check}` imports from
   `src/features/*/model/**` so this can't regress; the cross-feature
   patterns from the parent block are duplicated (with a comment) because
   ESLint flat config replaces — not merges — `no-restricted-imports`.

2. StandingsPage was 423 lines of orchestration + inline composed views.
   Extracted `useStandingsScreen` (`features/scoring/model`) for the hook
   wiring, and `StandingsTourView` / `StandingsShowOrPoolView`
   (`features/scoring/ui`) for the two presentational composition
   subviews. Page is now 41 lines of pure routing/composition.

No behavioral changes. Lint, vitest (122/122), verify:dashboard-meta,
verify:dashboard-ui, and vite build all green locally.

Closes #288

Made-with: Cursor
