+++ src/features/scoring/index.js
@@ -6,6 +6,7 @@ export { default as Leaderboard } from './ui/Leaderboard';
 export { useDisplayedPicks } from './model/useDisplayedPicks';
 export { useStandings } from './model/useStandings';
 export { useStandingsLeaderboardView } from './model/useStandingsLeaderboardView';
+export { useStandingsScreen } from './model/useStandingsScreen';
 export {
   DEFAULT_STANDINGS_VIEW,
   STANDINGS_VIEWS,
@@ -20,6 +21,8 @@ export {
 export { default as StandingsActiveShowCard } from './ui/StandingsActiveShowCard';
 export { default as StandingsBannerWaitingSetlist } from './ui/StandingsBannerWaitingSetlist';
 export { default as StandingsPoolPicker } from './ui/StandingsPoolPicker';
+export { default as StandingsShowOrPoolView } from './ui/StandingsShowOrPoolView';
+export { default as StandingsTourView } from './ui/StandingsTourView';
 export { default as StandingsViewToggle } from './ui/StandingsViewToggle';
 export { default as StandingsWinnerOfTheNightBanner } from './ui/StandingsWinnerOfTheNightBanner';
 export { default as TourStandingsSection } from './ui/TourStandingsSection';
